+++ 84 Super Star Trek/csharp/Objects/Klingon.cs
@@ -5,7 +5,7 @@ namespace SuperStarTrek.Objects
 {
     internal class Klingon
     {
-        private double _energy;
+        private float _energy;
         private Coordinates _sector;
         private readonly Random _random;
 
@@ -13,14 +13,14 @@ namespace SuperStarTrek.Objects
         {
             _sector = sector;
             _random = random;
-            _energy = _random.GetDouble(100, 300);
+            _energy = _random.GetFloat(100, 300);
         }
 
         public override string ToString() => "+K+";
 
         public CommandResult FireOn(Enterprise enterprise)
         {
-            var attackStrength = _random.GetDouble();
+            var attackStrength = _random.GetFloat();
             var hitStrength = (int)(_energy * (2 + attackStrength) / _sector.GetDistanceTo(enterprise.Sector));
             _energy /= 3 + attackStrength;
